Extract commit hash loader
