+++ service-worker.js
@@ -67,6 +67,7 @@ const APP_SHELL = [
   '/js/dashboard-widget-renderers.js',
   '/js/chart-card-recs.js',
   '/js/category-glyphs.js',
+  '/js/commit-hash.js',
   '/js/marker-detail-modal.js',
   '/js/light-conditions-now.js',
   '/js/light-sessions-view.js',
